Fix critical issue that caused the AI to store excessive hammers
This issue could occur if the AI was unable to supply non-hammer yields in order to complete the building.

It's been observed for one particular case, that 891 hammers were banked for a build whose cost was about 100. Since overflow is capped at the cost of the current building, this
represents a significant economic opportunity loss