Added script to configure project file to be able to start the mod with(out) the debugger attached
This script will take care of path configuration, something which is otherwise a hassle.
The debug menu "start debugging" and "start without debugging" now works as expected.
MSVC executes the makefile each time it starts the game, making sure the newest changes are included.
This includes updates to the autogenerated headers based on xml data.